v2 prep: route history_processors deprecation through PydanticAIDeprecationWarning
- Swap bare DeprecationWarning for PydanticAIDeprecationWarning (subclass of
  UserWarning, visible by default; matches the convention picked in #5338).
- Reword the warning to embed the before/after migration shape inline, per
  V2-RULES rule 9. Now reads:
  `Agent(history_processors=[fn, ...])` is deprecated and will be removed in
  v2.0. Replace with `Agent(capabilities=[ProcessHistory(fn), ...])`.
- Move the kwarg-deprecation test to tests/v2/ per V2-RULES rule 6.
- Migrate the internal test_temporal.py caller off the kwarg.
