+++ pydantic_ai_slim/pydantic_ai/.agents/skills/building-pydantic-ai-agents/SKILL.md
@@ -254,7 +254,7 @@ These are mistakes agents commonly make with Pydantic AI. Getting these wrong pr
 - **`TestModel` requires `agent.override()`**: Don't set `agent.model` directly. Always use the context manager: `with agent.override(model=TestModel()):`.
 - **`str` in output_type allows plain text to end the run**: If your union includes `str` (or no `output_type` is set), the model can return plain text instead of structured output. Omit `str` from the union to force tool-based output.
 - **Hook decorator names on `.on` don't repeat `on_`**: Use `hooks.on.run_error` and `hooks.on.model_request_error` — not `hooks.on.on_run_error`.
-- **`history_processors` is deprecated; use `capabilities=[ProcessHistory(p), ...]`**. The kwarg still works in 1.x but emits a `DeprecationWarning` and will be removed in v2.
+- **`history_processors` is deprecated; use `capabilities=[ProcessHistory(p), ...]`**. The kwarg still works in 1.x but emits a `PydanticAIDeprecationWarning` and will be removed in v2.
 
 ## Task-Family References
